+++ include/nanobind/stl/unique_ptr.h
@@ -32,10 +32,8 @@ template <typename T> struct deleter {
     void operator()(void *p) noexcept {
         if (o) {
             // Don't run the deleter if the interpreter has been shut down
-            if (!NB_CALL(is_alive)())
-                return;
-            gil_scoped_acquire guard;
-            Py_DECREF(o);
+            if (detail::cleanup_guard guard{})
+                Py_DECREF(o);
         } else {
             delete (T *) p;
         }