+++ tests/test_intrusive.cpp
@@ -33,11 +33,13 @@ NB_MODULE(test_intrusive_ext, m) {
     nb::intrusive_init(
         [](PyObject *o) noexcept {
             nb::gil_scoped_acquire guard;
-            Py_INCREF(o);
+            if (guard.is_valid())
+                Py_INCREF(o);
         },
         [](PyObject *o) noexcept {
             nb::gil_scoped_acquire guard;
-            Py_DECREF(o);
+            if (guard.is_valid())
+                Py_DECREF(o);
         });
 
     nb::class_<nb::intrusive_base>(